releasetools: Store the build.prop file in the OTA zip
This file is often used to read information about the update contained
in the OTA. Place it in the update so it can be used by scripts.

The file is not added to the updater-script, so it will not be placed
onto the actual system.

Change-Id: I88044796cbe8f199ca02df2840fd944cba2c73fa
(cherry picked from commit 0754268ab920ad5337c90b271f34d0ca6959e78d)